+++ src/custom-event-target.ts
@@ -20,26 +20,6 @@ type EventOptionsType<T> = T extends new (
   ? O
   : never;
 
-export function asTypedEvent<
-  const E extends Event,
-  const A = EventInit
->(event: { new (type: string, options?: A): E }) {
-  return event as {
-    new <
-      T extends EventTarget,
-      const U,
-      const O extends EventOptionsType<typeof event>
-    >(
-      type: U,
-      options?: O
-    ): BaseTypedEvent<T, EventInstanceType<typeof event, O>, U>;
-    prototype: BaseTypedEvent<EventTarget, E, string>;
-  };
-}
-
-export type TypedEvent<T extends EventTarget, U> = BaseTypedEvent<T, Event, U>;
-export const TypedEvent = asTypedEvent(Event);
-
 export type TypedCustomEvent<
   D,
   T extends EventTarget,